requirements.txt
scipy
scikit-learn
numpy
Pillow
opencv_python
tensorboardX
ftfy
matplotlib
regex
imageio
scikit-image
blobfile

diffusers
transformers
segmoe
sentencepiece
beautifulsoup4
ftfy
huggingface-hub
tqdm
dominate

